refactor: Rename utils/ to util/
