+++ src/winter/util/Utility.java
@@ -1,4 +1,4 @@
-package winter.utils;
+package winter.util;
 
 public class Utility {
     protected Utility() throws IllegalStateException {
